+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## 0.9.0 - 2021-01-31
 ### Added
 - Builder type implements `Clone` when all set fields support clone.
+
+### Fixed
 - Fix code generation for raw identifiers.
 
 ## 0.8.0 - 2020-12-06